+++ src/Exception/RadExceptionInterface.php
@@ -1,7 +1,0 @@
-<?php declare(strict_types=1);
-
-namespace Becklyn\Rad\Exception;
-
-interface RadExceptionInterface extends \Throwable
-{
-}