+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 *   (internal) Renamed Composer package from `becklyn/rad-bundle` to `becklyn/rad`. See README for more information.
 *   (bc) Rename namespace `Becklyn\RadBundle` to `Becklyn\Rad`.
 *   (improvement) Move `isNew()` to `EntityInterface` as well (is automatically implemented by `IdTrait`).
+*   (internal) `RadException` is now an interface.
+*   (bc) Remove `RadExceptionInterface`.
 
 
 7.14.1